--- react-chat/src/App.module.css
@@ -1,13 +1,0 @@
-.allChats {
-  width: 500px;
-  min-height: 100vh;
-  background-color: #f5f5f5;
-}
-
-.chatPage {
-      min-height: 100vh;
-      display: flex;
-      flex-direction: column;
-      background-color: #fff9f3;
-      flex: 1 0 auto;
-}
