+++ src/components/FloatButton/FloatButton.module.scss
@@ -1,0 +1,11 @@
+.floatButton {
+    position: absolute;
+    box-shadow: 2px 2px #c6c3c7;
+    width: 70px;
+    height: 70px;
+    border-radius: 50%;
+    border: 0;
+    background-color: #f9d667;
+    bottom: 25px;
+    left: 25px;
+}
