+++ src/components/FloatButton/index.js
@@ -1,0 +1,1 @@
+export { FloatButton } from './FloatButton';
