+++ src/components/ProfileBody/index.js
@@ -1,0 +1,1 @@
+export { ProfileBody } from './ProfileBody';
