+++ src/components/ProfileHeader/ProfileHeader.module.scss
@@ -1,0 +1,38 @@
+header {
+    display: flex;
+    height: 80px;
+    min-width: 360px;
+    background-color: #8e24aa;
+}
+
+.backButton,
+.doneButton {
+    background-color: #8e24aa;
+    color: white;
+    border: 0;
+    flex-basis: 60px;
+    transition: background-color 0.1s, color 0.1s;
+}
+
+.backButton:hover,
+.doneButton:hover {
+   background-color: #d37ceb;
+   color: black;
+}
+
+.userName {
+    font-weight: bold;
+    align-items: center;
+    color: white;
+    padding-left: 20px;
+    flex-grow: 2;
+    align-self: center;
+}
+
+@media screen and (max-width: 500px) {
+
+}
+
+@media screen and (min-width: 500px) {
+
+}
