+++ src/components/ProfileHeader/index.js
@@ -1,0 +1,1 @@
+export { ProfileHeader } from './ProfileHeader';
