+++ src/pages/ChatPage/ChatPage.module.scss
@@ -1,0 +1,7 @@
+.chatPage {
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    background-color: #fff9f3;
+    flex: 1 0 auto;
+  }
