+++ src/pages/ProfilePage/ProfilePage.jsx
@@ -1,0 +1,15 @@
+import { ProfileHeader } from "../../components/ProfileHeader";
+import { ProfileBody } from "../../components/ProfileBody";
+
+import styles from "./ProfilePage.module.scss";
+
+function ProfilePage() {
+  return (
+    <aside className={styles.profilePage}>
+      <ProfileHeader />
+      <ProfileBody />
+    </aside>
+  );
+}
+
+export { ProfilePage };
