+++ src/pages/ProfilePage/ProfilePage.module.scss
@@ -1,0 +1,7 @@
+.profilePage {
+    display: flex;
+    flex-direction: column;
+    width: 500px;
+    min-height: 100vh;
+    background-color: #ffffff;
+  }
