+++ src/pages/ProfilePage/index.js
@@ -1,0 +1,1 @@
+export { ProfilePage } from './ProfilePage';
