+++ src/pages/SidebarPage/SidebarPage.module.scss
@@ -1,0 +1,5 @@
+.allChats {
+    width: 500px;
+    min-height: 100vh;
+    background-color: #f5f5f5;
+  }
